+++ public/css/add-book.css
@@ -1,0 +1,132 @@
+.form-container-addbook {
+    max-width: 650px;
+    width: 90%; 
+    margin: 0 auto;
+    padding: 40px;
+    background-color: white;
+    border: 2px solid black;
+    border-radius: 12px;
+    box-shadow: 0 10px 20px rgba(0, 0, 0, 0.2);
+    display: none;
+    position: fixed;
+    top: 50%;
+    left: 50%;
+    transform: translate(-50%, -50%);
+    z-index: 1000;
+    text-align: center;
+}
+
+.form-header {
+    margin-bottom: 20px;
+}
+
+.form-content {
+    display: flex;
+    gap: 60px;
+}
+
+.image-preview-container-addbook {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    width: 200px;
+}
+
+.image-preview-addbook {
+    width: 100%;
+    height: 300px;
+    border: 2px solid black;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    background-color: #f0f0f0;
+    padding: 10px;
+    overflow: hidden;
+}
+
+.image-preview-addbook img {
+    width: 100%;
+    height: 100%;
+    object-fit: cover; /* Ensures the image fully fills the box */
+}
+
+.file-input.addbook {
+    margin-top: 10px;
+    width: 100%;
+}
+
+.form-fields-addbook {
+    flex: .85;
+}
+
+.form-group {
+    margin-bottom: 15px;
+    text-align: left;
+}
+
+.form-group label {
+    font-weight: bold;
+    margin-bottom: 5px;
+    display: block;
+}
+
+.form-group input{
+    width: 100%;
+    padding: 12px;
+    border: 1px solid black;
+    border-radius: 8px;
+
+}
+
+.form-group select{
+    background-color: white;
+    font-size: 20px;
+    width: 105%;
+}
+
+.button-container {
+    text-align: center;
+    margin-top: -90px;
+}
+
+.submit-btn {
+    width: 200px;
+    padding: 14px;
+    background-color: #2e7d32;
+    color: white;
+    border: 2px solid #2e7d32;
+    border-radius: 8px;
+    cursor: pointer;
+    font-weight: bold;
+    transition: background-color 0.3s ease, color 0.3s ease;
+}
+
+.submit-btn:hover {
+    background-color: #1b5e20;
+    border-color: #1b5e20;
+    color: white;
+}
+
+.overlay {
+    position: fixed;
+    top: 0;
+    left: 0;
+    width: 100%;
+    height: 100%;
+    background-color: rgba(0, 0, 0, 0.5);
+    z-index: 500;
+    display: none;
+}
+
+.container {
+    /* Optional: to center the content */
+    display: flex;
+    justify-content: center;
+    align-items: center;
+}
+
+.sized-box {
+    width: 100px; /* Set your desired width */
+    height: 20px; /* Set your desired height */
+    background-color: #dddddd00; /* Optional: background color for visibility */
+}
