+++ public/index.html
@@ -9,60 +9,89 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>BookTrack - Add Book</title>
     <link rel="stylesheet" href="css/style.css">
+    <link rel="stylesheet" href="css/add-book.css">
 </head>
 <body>
     <div class="header">
         <h1>BookTrack</h1>
         <div class="button-container">
             <button class="add-book-btn" onclick="openForm()">Add Book</button>
-            <button class="open-library-btn" onclick="openModal()">Open Library Form</button>
+            <button class="open-library-btn" onclick="openModal()">Add borrow book Form</button>
         </div>
     </div>
     <!-- Overlay to display a semi-transparent background when the form is open, closing form on click -->
     <div class="overlay" id="overlay" onclick="closeForm()"></div>
 
+    <!-- START FOR ADD-BOOK FUNCTION -->
+
     <!-- Container for the 'Add Book' form -->
-    <div class="form-container" id="formContainer">
-        <h2>Add a New Book</h2>
-        <form id="bookForm" enctype="multipart/form-data">
-            <div class="form-group">
-                <label for="title">Title:</label>
-                <input type="text" id="title" name="title">
+    <div class="form-container-addbook" id="formContainer">
+        <!-- Header for the form -->
+        <h2 class="form-header">Add a New Book</h2>
+        <div class="form-content">
+            <!-- Container to preview the selected image for the book -->
+            <div class="image-preview-container-addbook">
+                <div class="image-preview-addbook" id="imagePreview">
+                    <span>No Image Selected</span>
+                </div>
+                <div class="container">
+                    <!-- Spacing box to maintain consistent layout -->
+                    <div class="sized-box"></div>
+                </div>
+                <!-- Input field for selecting an image file -->
+                <input type="file" id="image" name="image" accept="image/*" class="file-input-addbook">
+                
             </div>
-            <div class="form-group">
-                <label for="author">Author:</label>
-                <input type="text" id="author" name="author">
+            <!-- Container for the book input fields in the form -->
+            <div class="form-fields-addbook">
+                <form id="bookForm" enctype="multipart/form-data">
+                    <!-- Form group for book title input -->
+                    <div class="form-group">
+                        <label for="title">Title:</label>
+                        <input type="text" id="title" name="title">
+                    </div>
+                    <!-- Form group for book author input -->
+                    <div class="form-group">
+                        <label for="author">Author:</label>
+                        <input type="text" id="author" name="author">
+                    </div>
+                    <!-- Form group for ISBN input -->
+                    <div class="form-group">
+                        <label for="isbn">ISBN:</label>
+                        <input type="text" id="isbn" name="isbn">
+                    </div>
+                    <!-- Form group for genre selection -->
+                    <div class="form-group">
+                        <label for="genre">Genre:</label>
+                        <select id="genre" name="genre">
+                            <option value="" disabled selected>Select Genre</option>
+                            <option value="Fiction">Fiction</option>
+                            <option value="Non-Fiction">Non-Fiction</option>
+                            <option value="Mystery">Mystery</option>
+                            <option value="Science Fiction">Science Fiction</option>
+                            <option value="Fantasy">Fantasy</option>
+                        </select>
+                    </div>
+                    <!-- Form group for available copies input -->
+                    <div class="form-group">
+                        <label for="copies">Available Copies:</label>
+                        <input type="number" id="copies" name="copies" min="1">
+                    </div>
+                </form>
             </div>
-            <div class="form-group">
-                <label for="isbn">ISBN:</label>
-                <input type="text" id="isbn" name="isbn">
-            </div>
-            <!-- Form group for genre selection dropdown -->
-            <div class="form-group">
-                <label for="genre">Genre:</label>
-                <select id="genre" name="genre">
-                    <option value="" disabled selected>Select Genre</option>
-                    <option value="Fiction">Fiction</option>
-                    <option value="Non-Fiction">Non-Fiction</option>
-                    <option value="Mystery">Mystery</option>
-                    <option value="Science Fiction">Science Fiction</option>
-                    <option value="Fantasy">Fantasy</option>
-                </select>
-            </div>
-            <!-- Form group for inputting available copies -->
-            <div class="form-group">
-                <label for="copies">Available Copies:</label>
-                <input type="number" id="copies" name="copies" min="1">
-            </div>
-            <!-- Form group for image file upload -->
-            <div class="form-group">
-                <label for="image">Upload Image:</label>
-                <input type="file" id="image" name="image" accept="image/*">
-            </div>
-            <!-- Submit button to add the book -->
-            <button type="submit" class="submit-btn">Add Book</button>
-        </form>
+        </div>
+
+    <!-- Submit button positioned outside the form but within the form container -->
+    <div class="button-container">
+            <button type="submit" form="bookForm" class="submit-btn">Add Book</button>
+        </div>
     </div>
+    
+    <!-- Link to the external JavaScript file for handling 'Add Book' form interactions -->
+    <script src="js/add-book.js"></script>
+
+    <!-- END FOR ADD-BOOK FUNCTION -->
+
 
  
 
@@ -70,7 +99,7 @@
     <div class="modal-overlay" id="modalOverlay">
     <div class="modal-container">
         <span id="closeBtn" class="close-btn" onclick="closeModal()">&times;</span>
-        <h1 id="formTitle" class="form-title">Library Book Borrowing Form</h1>
+        <h1 id="formTitle" class="form-title">Add Library Book Borrowing  </h1>
         <form id="transactionForm">
             <label for="book_id">Book ID</label>
             <input type="text" id="book_id" name="book_id" placeholder="Enter Book ID" required>
@@ -94,8 +123,7 @@
 
 
     <!-- Link to the external JavaScript file for form functionality -->
-    <script src="js/Add-Book.js"></script>
-<script src="js/Add_Transaction.js"></script>
+<script src="js/add-transaction.js"></script>
 <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.0.0-beta3/css/all.min.css">
 
 
@@ -127,8 +155,8 @@
  
 </body>
 
-<body>
-
+<body>  
+    <!-- Container for the edit book form, initially hidden -->
     <div class="form-container" id="editFormContainer" style="display: none;">
         <h2 style="text-align: center;">Edit Book Details</h2>
         <div class="form-content">
@@ -178,8 +206,8 @@
             </div>
         </div>
     </div>
- <script src="js/updateBook.js"></script>
-
+    
+<script src="js/update-book.js"></script>
 </body>
 
 </html>
